Fixed bug in code revealed by live testing on dev instance
After pushing the code to the dev instance on aws, found that the modifications added for mocking during tests actually only worked during tests, and would not work at all in a live environment. This commit fixes those issues and has been tested again on the dev instance.

Part of #10